Move non-bokeh web content to http directory
