+++ distributed/dashboard/utils.py
@@ -1,10 +1,8 @@
 from distutils.version import LooseVersion
-import os
 from numbers import Number
 
 import bokeh
 from bokeh.io import curdoc
-from tornado import web
 from tlz import partition
 from tlz.curried import first
 
@@ -15,7 +13,6 @@ except ImportError:
 
 
 BOKEH_VERSION = LooseVersion(bokeh.__version__)
-dirname = os.path.dirname(__file__)
 
 
 PROFILING = False
@@ -43,23 +40,6 @@ def parse_args(args):
 def transpose(lod):
     keys = list(lod[0].keys())
     return {k: [d[k] for d in lod] for k in keys}
-
-
-class RequestHandler(web.RequestHandler):
-    def initialize(self, server=None, extra=None):
-        self.server = server
-        self.extra = extra or {}
-
-    def get_template_path(self):
-        return os.path.join(dirname, "templates")
-
-
-def redirect(path):
-    class Redirect(RequestHandler):
-        def get(self):
-            self.redirect(path)
-
-    return Redirect
 
 
 @without_property_validation
