+++ distributed/http/templates/simple.html
