+++ distributed/http/templates/worker-table.html
