+++ distributed/http/templates/worker.html
