+++ setup.py
@@ -30,7 +30,7 @@ setup(
     license="BSD",
     package_data={
         "": ["templates/index.html", "template.html"],
-        "distributed": ["dashboard/templates/*.html"],
+        "distributed": ["http/templates/*.html"],
     },
     include_package_data=True,
     install_requires=install_requires,
